Broke GameTree into Interface and Implementation
